workflow_service/requirements.txt
celery==5.4.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
redis==5.2.1
aiohttp==3.11.11
asyncio==3.4.3
requests==2.32.3